requirements.txt
pipecat-ai
twilio
supabase
google-api-python-client
google-auth-oauthlib
openai
python-dotenv
deepgram-sdk
elevenlabs
fastapi
uvicorn
python-multipart
